Create Modern and Competitive Quiz Generator
The idea for the project is named as Quiz Generator which is an application developed using Python that can generate quiz of various subjects such as Data Structures and Algorithms, Design and Analysis of Algorithms, Computer Organization and Architecture, and Physics. This application stands out from conventional quizzes by offering four distinct types of quizzes:  Easy, Medium, Difficult, and Competition.
The possibilities are as follows: user can choose the level of difficulty beginning from Easy which enables certain features like skip and go back and Medium. Still, the Difficult quiz will set a time constraint and a non-returning/ non-skip feature to increase the level of complication for the user.
Further, the Competition quiz mode also brings another interactive twist where up to 5 players can play at the same time. The quiz will commence immediately all participants are set and this will be a total of 10 questions, from basic to advanced level. This one is point based, where users quickly responding are rewarded; the more apt a user is in responding, the more points he gets. For the first 5 questions only the top 3 participants will be asked the remaining 5 questions and the bottom 2 will be asked nothing. The participants will have a leaderboard where results will change with each question, therefore making the competition like environment.